faceAcecomV1/requirements.txt
altair==4.1.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
base58==2.0.1
bleach==3.1.0
blinker==1.4
boto3==1.15.17
botocore==1.18.17
cachetools==4.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
configparser==4.0.2
crayons==0.3.0
cycler==0.10.0
dash==0.23.1
dash-core-components==0.26.0
dash-html-components==0.11.0
dash-renderer==1.2.4
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
enum-compat==0.0.3
enum34==1.1.6
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.9
google-pasta==0.1.7
h5py==2.10.0
html5lib==0.9999999
idna==2.8
imageio==2.6.1
importlib-metadata==1.3.0
imutils==0.5.3
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.0.2
nb-conda==2.2.1
nb-conda-kernels==2.2.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.16.1
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pickleshare==0.7.5
Pillow==6.2.1
plotly==3.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.13.0
pyarrow==1.0.1
PyAudio==0.2.11
pydeck==0.5.0b1
Pygments==2.5.2
pynput==1.6.6
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.0
qtconsole==4.6.0
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
scikit-learn==0.19.1
scipy==1.2.2
selenium==3.141.0
Send2Trash==1.5.0
six==1.13.0
smmap==3.0.4
streamlit==0.69.1
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tflearn==0.3.2
toml==0.10.1
toolz==0.11.1
tornado==6.0.3
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.8
validators==0.18.1
watchdog==0.10.3
wcwidth==0.1.7
webdriver-manager==2.3.0
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
